+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added test case for replace all
 - Added build_simple.bat for windows
 - Added WsjcppCore::makeDirsPath
+- Added WsjcppCore::extractDirpath
 
 ### Changed